requirements.txt
djangorestframework-datatables
Django
djangorestframework
django-select2